Rename `SimpleCache` to `CacheStore`
