+++ packages/drift/src/client/MockClient.ts
@@ -1,26 +1,26 @@
 import { MockAdapter } from "src/adapter/MockAdapter";
 import type { OxAdapterConfig } from "src/adapter/OxAdapter";
-import type { LruSimpleCache } from "src/cache/LruSimpleCache";
-import type { SimpleCache } from "src/cache/types";
 import {
   type Client,
   type ClientConfig,
   createClient,
 } from "src/client/Client";
+import type { LruStore } from "src/store/LruStore";
+import type { CacheStore } from "src/store/types";
 
 export type MockClient<
   TAdapter extends MockAdapter = MockAdapter,
-  TCache extends SimpleCache = SimpleCache,
+  TCache extends CacheStore = CacheStore,
 > = Client<TAdapter, TCache>;
 
 export type MockClientConfig<
   TAdapter extends MockAdapter = MockAdapter,
-  TCache extends SimpleCache = SimpleCache,
+  TCache extends CacheStore = CacheStore,
 > = Partial<Omit<ClientConfig<TAdapter, TCache>, keyof OxAdapterConfig>>;
 
 export function createMockClient<
   TAdapter extends MockAdapter = MockAdapter,
-  TCache extends SimpleCache = LruSimpleCache,
+  TCache extends CacheStore = LruStore,
 >({
   adapter = new MockAdapter() as TAdapter,
   chainId,
