+++ packages/drift/src/client/MockDrift.ts
@@ -1,15 +1,15 @@
 import type { MockAdapter } from "src/adapter/MockAdapter";
 import type { Abi } from "src/adapter/types/Abi";
 import type { ContractParams } from "src/adapter/types/Contract";
-import type { LruSimpleCache } from "src/cache/LruSimpleCache";
-import type { SimpleCache } from "src/cache/types";
 import type { Client } from "src/client/Client";
 import { type MockClientConfig, createMockClient } from "src/client/MockClient";
 import { MockContract } from "src/client/contract/MockContract";
+import type { LruStore } from "src/store/LruStore";
+import type { CacheStore } from "src/store/types";
 
 export type MockDrift<
   TAdapter extends MockAdapter = MockAdapter,
-  TCache extends SimpleCache = SimpleCache,
+  TCache extends CacheStore = CacheStore,
 > = Client<
   TAdapter,
   TCache,
@@ -28,7 +28,7 @@ export type MockDrift<
 
 export function createMockDrift<
   TAdapter extends MockAdapter = MockAdapter,
-  TCache extends SimpleCache = LruSimpleCache,
+  TCache extends CacheStore = LruStore,
 >(
   config: MockClientConfig<TAdapter, TCache> = {},
 ): MockDrift<TAdapter, TCache> {
